requirements.txt
flask==3.0.2
requests==2.31.0
python-dotenv==1.0.1
google-search-results==2.4.2
reportlab==4.1.0
gunicorn==21.2.0
pymongo==4.6.1
langchain==0.1.16
faiss-cpu==1.7.4
sentence-transformers==2.6.1
beautifulsoup4==4.12.3